Extracing the Generate Decorator to a class
